+++ src/developer-platform/service-registration.js
@@ -1,6 +1,7 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
 import { DeveloperPlatformOverviewDescriptor } from './overview/developer-platform-overview-descriptor.js';
 import { DeveloperExperienceDescriptor } from './experience/developer-experience-descriptor.js';
+import { PlatformCapabilityModelDescriptor } from './capability-model/platform-capability-model-descriptor.js';
 
 export function addDeveloperPlatform(services) {
   if (!(services instanceof ServiceCollection)) {
@@ -9,6 +10,7 @@ export function addDeveloperPlatform(services) {
 
   services.registerSingleton('DeveloperPlatformOverviewDescriptor', () => new DeveloperPlatformOverviewDescriptor());
   services.registerSingleton('DeveloperExperienceDescriptor', () => new DeveloperExperienceDescriptor());
+  services.registerSingleton('PlatformCapabilityModelDescriptor', () => new PlatformCapabilityModelDescriptor());
 
   return services;
 }